refactor: Recipeのtypeをリファクタリング
