+++ frontend/src/pages-component/RecipeDetail/RecipeDetail/RecipeDetail.stories.ts
@@ -1,6 +1,6 @@
 import { type Meta, type StoryObj } from "@storybook/react";
 import { SPStory } from "@/lib/storybook/SPStory";
-import { recipeDetailArgs } from "@/lib/storybook/recipeArgs";
+import { imageUrl, recipe, name } from "@/lib/storybook/recipeArgs";
 import { RecipeDetail } from "@/pages-component/RecipeDetail/RecipeDetail/RecipeDetail";
 
 const meta: Meta<typeof RecipeDetail> = {
@@ -12,6 +12,12 @@ const meta: Meta<typeof RecipeDetail> = {
 export default meta;
 
 type Story = StoryObj<typeof RecipeDetail>;
+
+const recipeDetailArgs = {
+  ...name,
+  ...imageUrl,
+  recipe,
+};
 
 export const Default: Story = {
   parameters: {
