+++ frontend/src/pages-component/RecipeDetail/RecipeImage/RecipeImage.stories.ts
@@ -1,6 +1,6 @@
 import { type StoryObj, type Meta } from "@storybook/react";
 import { SPStory } from "@/lib/storybook/SPStory";
-import { recipeImageArgs } from "@/lib/storybook/recipeArgs";
+import { imageUrl, name } from "@/lib/storybook/recipeArgs";
 import { RecipeImage } from "@/pages-component/RecipeDetail/RecipeImage/RecipeImage";
 
 const meta: Meta<typeof RecipeImage> = {
@@ -12,6 +12,11 @@ const meta: Meta<typeof RecipeImage> = {
 export default meta;
 
 type Story = StoryObj<typeof RecipeImage>;
+
+const recipeImageArgs = {
+  ...name,
+  ...imageUrl,
+};
 
 export const Default: Story = {
   parameters: {
