+++ frontend/src/pages-component/RecipeRecommender/RecipeCard/RecipeCard.stories.ts
@@ -1,6 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 import { SPStory } from "@/lib/storybook/SPStory";
-import { recipeCardArgs } from "@/lib/storybook/recipeArgs";
+import { imageUrl, recipe, name } from "@/lib/storybook/recipeArgs";
 import { RecipeCard } from "./RecipeCard";
 
 const meta: Meta<typeof RecipeCard> = {
@@ -15,6 +15,14 @@ const meta: Meta<typeof RecipeCard> = {
 export default meta;
 
 type Story = StoryObj<typeof RecipeCard>;
+
+const recipeCardArgs = {
+  ...name,
+  ...imageUrl,
+  recipe,
+  // eslint-disable-next-line @typescript-eslint/no-empty-function, @typescript-eslint/explicit-module-boundary-types
+  onClick: () => {},
+};
 
 export const Default: Story = {
   parameters: {
