+++ frontend/src/pages-component/RecipeRecommender/RecipeRecommender/RecipeRecommender.stories.ts
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
 import { SPStory } from "@/lib/storybook/SPStory";
-import { recipeCardArgs } from "@/lib/storybook/recipeArgs";
+import { imageUrl, recipe, name } from "@/lib/storybook/recipeArgs";
 import { RecipeRecommender } from "@/pages-component/RecipeRecommender/RecipeRecommender/RecipeRecommender";
 
 const meta: Meta<typeof RecipeRecommender> = {
@@ -14,12 +14,20 @@ export default meta;
 
 type Story = StoryObj<typeof RecipeRecommender>;
 
+const RecipeRecommenderArgs = {
+  ...name,
+  ...imageUrl,
+  recipe,
+  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types, @typescript-eslint/no-empty-function
+  showDetailPage: () => {},
+};
+
 export const Default: Story = {
   parameters: {
     ...SPStory.parameters,
   },
-  args: recipeCardArgs,
+  args: RecipeRecommenderArgs,
 };
 export const PCDefault: Story = {
-  args: recipeCardArgs,
+  args: RecipeRecommenderArgs,
 };
