--- frontend/src/types/index.ts
@@ -1,7 +1,0 @@
-import { type Url } from "url";
-
-export type Recipe = {
-  name: string;
-  imageUrl: Url;
-  text: string;
-};
